backend/requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-cors==0.8.1
aiosignal==1.3.2
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
cssselect==1.3.0
decorator==4.4.2
dnspython==2.7.0
ensta==5.2.9
fake-http-header==0.3.5
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
lxml==5.3.2
motor==3.7.0
moviepy==1.0.1
multidict==6.4.3
ntplib==0.4.0
numpy==2.2.4
pillow==10.4.0
playwright==1.51.0
proglog==0.1.11
propcache==0.3.1
pycares==4.6.0
pycparser==2.22
pyee==12.1.1
pymongo==4.12.0
pyotp==2.9.0
pyquery==2.0.1
python-dotenv==1.1.0
requests==2.32.3
tf-playwright-stealth==1.1.2
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.19.0
